print sig and groundtruth

+++ pytorchtest.py
@@ -116,7 +116,7 @@ def train_model(model, criterion, optimizer, lr_scheduler):
                 optimizer.zero_grad()
                 # forward
                 outputs = model(inputs)
-                preds = torch.sigmoid(outputs.data).long()
+                preds = torch.sigmoid(outputs.data)
                 loss = criterion(outputs, labels)
                 # backward + optimize only if in training phase
                 if phase == 'train':
@@ -124,7 +124,7 @@ def train_model(model, criterion, optimizer, lr_scheduler):
                     optimizer.step()
                 # statistics
                 running_loss += loss.data[0]
-                running_corrects += torch.sum(preds == labels.data.long())
+                running_corrects += torch.sum(preds.long() == labels.data.long())
                 # next step
                 step += 1
             epoch_loss = running_loss / dset_sizes[phase]
